Added Autos and Fixed Shooter
inverted the second shooter motor for consistency and added an intake from Shooter command for picking up from source

added autos for 3, 4, and 5 note autos and options for not going to the subwoofer each time. Added a MessUp auto to shoot one and mess with the center notes to decrease chance of anyone getting them

Cleaned up code and removed unnecessary paths.